[BUG] second fix for the printf format warning
Fix 500b8f0349fb52678f5143c49f5a8be5c033a988 fixed the patch for the 64 bit
case but caused the opposite type issue to appear on 32 bit platforms. Cast
the difference and be done with it since gcc does not agree on type carrying
the difference between two pointers on 32 and 64 bit platforms.

+++ src/cfgparse.c
@@ -4085,7 +4085,7 @@ int readcfgfile(const char *file)
 			*line = '\0';
 
 			Alert("parsing [%s:%d]: line too long, truncating at word %d, position %ld: <%s>.\n",
-			      file, linenum, arg + 1, args[arg] - thisline + 1, args[arg]);
+			      file, linenum, arg + 1, (long)(args[arg] - thisline + 1), args[arg]);
 			err_code |= ERR_ALERT | ERR_FATAL;
 			args[arg] = line;
 		}
